features: time between SRD and addIceCandidate
this adds two features which measure the delay between
* SRD and addIceCandidate
* SLD and onicecandidate
I found a number of ice failures where there was a delay of >5 seconds between
SRD and ice candidate which caused firefox to fail. I also found some sessions
where there was a large delay between SLD and onicecandidate which would cause
Firefox to fail as well.