fix track extraction confusion when send and receive track have the same id
Summary: When the send and receive trackshave the same id which can happen in ios-ios cases we extracted the track stats solely on the track name. Which was mixing up send and receive stats, leading to the extraction of wrong values such as NaN for GoogFrameWidthReceived due to the presence of stats which did not contain a key.

Test Plan: - test with attached dump, check videoGoogFrameHeightReceivedMode is not NaN

Reviewers: pavel, gustavo

Reviewed By: pavel

Differential Revision: https://lifeonair.phacility.com/D25625