add iceconnectionstatecheckingbeforesrd feature
Summary:
adds a feature to detect a statechange in ice where it goes to checking before setRemoteDescription (or a ice candidate) is added. This can happen when we are the offerer and the peer sends us a stun ping. It has led to   https://bugs.chromium.org/p/chromium/issues/detail?id=959128#c65 so I am curious if there are other issues with those prflx candidates.

SQL:
alter table rtcstats add iceconnectionstatecheckingbeforesrd boolean

Test Plan:
- run node extract madness-newer.json
- observe feature being extract (as false)

Reviewers: gustavo

Reviewed By: gustavo

Differential Revision: https://lifeonair.phacility.com/D26883