add fallback identity information
if its not possible to pass extra things to the peerconnection constructor provide
a fallback way to set the client and conference ids. The rtcstats trace looks like this:

    ['identity', null, {user: 'something', conference: 'firechat'}]

where `user` is the same thing passed as rtcStatsClientIdentifier and `conference`
is the same thing passed as rtcStatsConferenceIdentifier in the non-fallback case.